Build Chair.py For IClone
This simple example demonstrates how to browse for two textures the chair and the rails for the chair. Then the chair is assembled and the pieces for the left rail are attached to from the leftRail, rightRail and backRail props. Finally the chair cushions are created, rotated, scaled and assembled into the form of a chair.

The final part attached the pieces to the parent, so you end up with a chair prop.